Separate controls and windows
