+++ OpenTabletDriver/Controls/LogView.xaml
@@ -3,7 +3,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-    x:Class="OpenTabletDriver.Views.LogView">
+    x:Class="OpenTabletDriver.Controls.LogView">
     <ListBox Items="{Binding Messages, Mode=TwoWay}" BorderThickness="0,1">
         <ListBox.ItemTemplate>
             <DataTemplate>
