+++ OpenTabletDriver/Controls/StatusBar.xaml
@@ -5,7 +5,7 @@
     mc:Ignorable="d"
     d:DesignWidth="800" d:DesignHeight="450"
     MinHeight="18"
-    x:Class="OpenTabletDriver.Views.StatusBar">
+    x:Class="OpenTabletDriver.Controls.StatusBar">
     <UserControl.Styles>
         <Style Selector="Border">
             <Setter Property="BorderBrush" Value="{DynamicResource ThemeBorderMidBrush}"/>
