+++ OpenTabletDriver/ViewLocator.cs
@@ -4,7 +4,6 @@
 using System;
 using Avalonia.Controls;
 using Avalonia.Controls.Templates;
-using OpenTabletDriver.ViewModels;
 
 namespace OpenTabletDriver
 {
