+++ OpenTabletDriver/Windows/About.xaml
@@ -3,7 +3,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="OpenTabletDriver.Views.About"
+        x:Class="OpenTabletDriver.Windows.About"
         Width="500" Height="300"
         Title="About OpenTabletDriver">
     <Window.Styles>
