+++ OpenTabletDriver/Windows/BindingConfig.xaml
@@ -3,7 +3,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-    x:Class="OpenTabletDriver.Views.BindingConfig"
+    x:Class="OpenTabletDriver.Windows.BindingConfig"
     Width="250" Height="250" CanResize="False"
     Title="Binding Configuration">
     <Window.Styles>
