+++ OpenTabletDriver/Windows/ConfigurationManager.xaml
@@ -1,15 +1,11 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="clr-namespace:OpenTabletDriver.ViewModels;assembly=OpenTabletDriver"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="OpenTabletDriver.Views.ConfigurationManager"
+        x:Class="OpenTabletDriver.Windows.ConfigurationManager"
         Width="1024" Height="768"
         Title="Configuration Manager">
-    <Design.DataContext>
-        <vm:ConfigurationManagerViewModel/>
-    </Design.DataContext>
     <Grid RowDefinitions="Auto,*">
         <Menu Grid.Row="0">
             <MenuItem Header="File">
