+++ OpenTabletDriver/Windows/TabletDebugger.xaml
@@ -2,9 +2,8 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:vm="clr-namespace:OpenTabletDriver.ViewModels;assembly=OpenTabletDriver"
     mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-    x:Class="OpenTabletDriver.Views.TabletDebugger"
+    x:Class="OpenTabletDriver.Windows.TabletDebugger"
     Width="800" Height="260"
     Title="Tablet Debugger">
     <Window.Styles>
