+++ OpenTabletDriver/Windows/TabletDebugger.xaml.cs
@@ -1,10 +1,17 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
+using TabletDriverLib;
+using TabletDriverPlugin.Tablet;
 
-namespace OpenTabletDriver.Views
+namespace OpenTabletDriver.Windows
 {
     public class TabletDebugger : Window
     {
+        public TabletDebugger(params DeviceReader<IDeviceReport>[] deviceReaders) : this()
+        {
+            DataContext = new TabletDebuggerViewModel(deviceReaders);
+        }
+        
         public TabletDebugger()
         {
             InitializeComponent();
